[#176575256] add cobadge api spec definitions move cobadge models to cobadge/configuration
