+++ package.json
@@ -11,7 +11,8 @@
   "io_bpd_winning_transactions": "https://raw.githubusercontent.com/pagopa/io-services-metadata/master/bonus/specs/bpd/winning_transactions.json",
   "pagopa_api": "https://raw.githubusercontent.com/pagopa/io-services-metadata/master/pagopa/pm/spec.json",
   "pagopa_api_walletv2": "https://raw.githubusercontent.com/pagopa/io-services-metadata/master/bonus/specs/bpd/pm/walletv2.json",
-  "model_pagopa_cobadge": "https://raw.githubusercontent.com/pagopa/io-services-metadata/master/pagopa/cobadge/abi_definitions.yml",
+  "pagopa_cobadge_configuration": "https://raw.githubusercontent.com/pagopa/io-services-metadata/master/pagopa/cobadge/abi_definitions.yml",
+  "pagopa_cobadge_api" : "https://raw.githubusercontent.com/pagopa/io-services-metadata/master/pagopa/cobadge/cobadge_1.6.4.json",
   "private": true,
   "scripts": {
     "start": "react-native start",
@@ -43,12 +44,13 @@
     "generate:bpd-award-periods": "rimraf definitions/bpd/award_periods && mkdir -p definitions/bpd/award_periods  && gen-api-models --api-spec $npm_package_io_bpd_award_periods --out-dir ./definitions/bpd/award_periods --no-strict --request-types --response-decoders",
     "generate:bpd-winning-transactions": "rimraf definitions/bpd/winning_transactions && mkdir -p definitions/bpd/winning_transactions  && gen-api-models --api-spec $npm_package_io_bpd_winning_transactions --out-dir ./definitions/bpd/winning_transactions --no-strict --request-types --response-decoders",
     "generate:pagopa-api-walletv2": "rimraf definitions/pagopa/walletv2 && mkdir -p definitions/pagopa/walletv2  && gen-api-models --api-spec $npm_package_pagopa_api_walletv2 --out-dir ./definitions/pagopa/walletv2 --no-strict --request-types --response-decoders",
-    "generate:pagopa-model-cobadge-definitions": "rimraf definitions/pagopa/cobadge && mkdir -p definitions/pagopa/cobadge && gen-api-models --api-spec $npm_package_model_pagopa_cobadge --out-dir ./definitions/pagopa/cobadge",
+    "generate:pagopa-cobadge-configuration": "rimraf definitions/pagopa/cobadge/configuration && mkdir -p definitions/pagopa/cobadge/configuration && gen-api-models --api-spec $npm_package_pagopa_cobadge_configuration --out-dir ./definitions/pagopa/cobadge/configuration",
+    "generate:pagopa-cobadge-api": "rimraf definitions/pagopa/cobadge && mkdir -p definitions/pagopa/cobadge && gen-api-models --api-spec $npm_package_pagopa_cobadge_api --out-dir ./definitions/pagopa/cobadge",
     "bundle:android-release": "node node_modules/react-native/local-cli/cli.js bundle --platform android --dev false --entry-file index.js --bundle-output android/app/src/main/assets/index.android.bundle --assets-dest android/app/src/main/res/",
     "generate:locales": "ts-node --skip-project -O '{\"lib\":[\"es2015\"]}' scripts/make-locales.ts",
     "generate:content-definitions": "rimraf definitions/content && mkdir -p definitions/content && gen-api-models --api-spec $npm_package_io_content_specs --out-dir ./definitions/content",
     "generate:bpd": "npm-run-all generate:bpd-citizen generate:bpd-payment generate:bpd-award-periods generate:bpd-winning-transactions",
-    "generate:all": "npm-run-all generate:mock-google-services-json generate:api-definitions generate:pagopa-api-definitions generate:pagopa-api-walletv2 generate:pagopa-model-cobadge-definitions generate:content-definitions generate:bonus-vacanze-definitions generate:bpd generate:locales",
+    "generate:all": "npm-run-all generate:mock-google-services-json generate:api-definitions generate:pagopa-api-definitions generate:pagopa-api-walletv2 generate:pagopa-cobadge-api generate:pagopa-cobadge-configuration generate:content-definitions generate:bonus-vacanze-definitions generate:bpd generate:locales",
     "locales_unused": "ts-node --skip-project -O '{\"lib\":[\"es2015\"]}' scripts/unused-locales.ts",
     "generate-myportal-readme": "jsdoc2md ts/utils/webviewScripts/*.js > MYPORTAL_README.md"
   },
